move cpio leader and load functions and data to a cpio module
