+++ src/main.rs
@@ -1,7 +1,6 @@
 use futures::stream::{FuturesOrdered, FuturesUnordered, TryStreamExt};
 use futures::StreamExt;
 use http::response::Builder;
-use lazy_static::lazy_static;
 use std::convert::TryInto;
 use std::ffi::OsString;
 use std::io;
@@ -23,11 +22,14 @@ extern crate log;
 mod cpio_cache;
 use crate::cpio_cache::CpioCache;
 
+mod cpio;
+use crate::cpio::{make_load_cpio, LEADER_CPIO_BYTES, LEADER_CPIO_LEN};
+
 mod options;
 use crate::options::Opt;
 
 mod files;
-use crate::files::{basename, open_file_stream};
+use crate::files::open_file_stream;
 
 mod hydra;
 
@@ -55,95 +57,6 @@ fn server_error() -> Rejection {
 fn feature_disabled(msg: &str) -> Rejection {
     warn!("Feature disabled: {}", msg);
     reject::not_found()
-}
-
-fn make_leader_cpio() -> std::io::Result<Vec<u8>> {
-    let mut leader_cpio = std::io::Cursor::new(vec![]);
-    cpio::write_cpio(
-        vec![
-            // mode for a directory: 0o40000 + 0o00xxx for its permission bits
-            // nlink for directories == the number of things in it plus 2 (., ..)
-            (
-                cpio::newc::Builder::new(".").mode(0o40755).nlink(3),
-                std::io::empty(),
-            ),
-            (
-                cpio::newc::Builder::new("nix").mode(0o40755).nlink(3),
-                std::io::empty(),
-            ),
-            (
-                cpio::newc::Builder::new("nix/store")
-                    .mode(0o40775)
-                    .nlink(2)
-                    .uid(0)
-                    .gid(30000),
-                std::io::empty(),
-            ),
-            (
-                cpio::newc::Builder::new("nix/.nix-netboot-serve-db")
-                    .mode(0o40755)
-                    .nlink(3),
-                std::io::empty(),
-            ),
-            (
-                cpio::newc::Builder::new("nix/.nix-netboot-serve-db/registration")
-                    .mode(0o40755)
-                    .nlink(2),
-                std::io::empty(),
-            ),
-        ]
-        .into_iter(),
-        &mut leader_cpio,
-    )?;
-
-    Ok(leader_cpio.into_inner())
-}
-
-fn make_load_cpio(paths: &Vec<PathBuf>) -> Result<Vec<u8>, LoadCpioError> {
-    let script = paths
-        .iter()
-        .map(|p| {
-            let mut line =
-                OsString::from("nix-store --load-db < /nix/.nix-netboot-serve-db/registration/");
-            line.push(basename(p).ok_or_else(|| LoadCpioError::NoBasename(p.to_path_buf()))?);
-            Ok(line)
-        })
-        .collect::<Result<Vec<OsString>, LoadCpioError>>()?
-        .into_iter()
-        .fold(OsString::from("#!/bin/sh"), |mut acc, line| {
-            acc.push("\n");
-            acc.push(line);
-            acc
-        });
-    let mut loader = std::io::Cursor::new(vec![]);
-    cpio::write_cpio(
-        vec![(
-            cpio::newc::Builder::new("nix/.nix-netboot-serve-db/register")
-                .mode(0o0100500)
-                .nlink(1),
-            std::io::Cursor::new(script.as_bytes()),
-        )]
-        .into_iter(),
-        &mut loader,
-    )
-    .map_err(LoadCpioError::Io)?;
-
-    Ok(loader.into_inner())
-}
-
-#[derive(Debug)]
-enum LoadCpioError {
-    Io(std::io::Error),
-    NoBasename(PathBuf),
-}
-
-lazy_static! {
-    static ref LEADER_CPIO_BYTES: Vec<u8> =
-        make_leader_cpio().expect("Failed to generate the leader CPIO.");
-    static ref LEADER_CPIO_LEN: u64 = LEADER_CPIO_BYTES
-        .len()
-        .try_into()
-        .expect("Failed to convert usize leader length to u64");
 }
 
 #[tokio::main]
